requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beanie==1.29.0
certifi==2025.4.26
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
toml==0.10.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
